+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 * Support user authentication provider type. ([#570](https://github.com/realm/realm-dart/pull/570))
 * Support user profile data. ([#570](https://github.com/realm/realm-dart/pull/570))
 * Support flexible synchronization. ([#496](https://github.com/realm/realm-dart/pull/496))
+* Added support for DateTime properties. ([#569](https://github.com/realm/realm-dart/pull/569))
 
 ### Fixed
 * Fixed an issue that would result in the wrong transaction being rolled back if you start a write transaction inside a write transaction. ([#442](https://github.com/realm/realm-dart/issues/442))